requirements.txt
requests==2.32.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Dev / test
pytest==8.2.2
pytest-cov==5.0.0
